pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI boilerplate for building production-ready microservices with clean architecture, SQLAlchemy, async support, and best practices out of the box."
authors = [
    {name = "Hector Lozada",email = "hlozadaccs@outlook.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pyotp (>=2.9.0,<3.0.0) ; python_version >= \"3.14\" and python_version < \"4.0\"",
    "qrcode[pil] (>=8.2,<9.0) ; python_version >= \"3.14\" and python_version < \"4.0\""
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.14.14,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "radon (>=6.0.1,<7.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "bandit[toml] (>=1.8.0,<2.0.0)",
    "safety (>=3.2.0,<4.0.0)",
    "interrogate (>=1.7.0,<2.0.0)",
    "vulture (>=2.14,<3.0)",
    "detect-secrets (>=1.5.0,<2.0.0)",
    "pyproject-fmt (>=2.12.1,<3.0.0)"
]

[tool.mypy]
python_version = "3.14"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.infrastructure.db.repository"
disable_error_code = ["attr-defined"]

[[tool.mypy.overrides]]
module = "app.main"
disable_error_code = ["method-assign"]

[[tool.mypy.overrides]]
module = "app.core.config"
disable_error_code = ["call-arg"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "C90", # mccabe (cyclomatic complexity)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from within except (intentional for API errors)
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"app/scripts/*.py" = ["C90"]  # Allow higher complexity in scripts

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.bandit]
exclude_dirs = ["tests", "alembic/versions"]
skips = ["B101"]  # Skip assert_used (common in tests)

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-nested-functions = true
fail-under = 40
exclude = ["tests", "alembic"]
verbose = 1

[tool.vulture]
min_confidence = 80
exclude = ["tests/", "alembic/"]
